+++ tests/unit/tf/models/test_retrieval.py
@@ -120,7 +120,7 @@ def test_two_tower_model_with_custom_options(
 ):
     from tensorflow.keras import regularizers
 
-    from merlin.models.tf.core.transformations import PopularityLogitsCorrection
+    from merlin.models.tf.transforms.bias import PopularityLogitsCorrection
     from merlin.models.utils import schema_utils
 
     data = ecommerce_data
